Finished Coding Navbar/Header
I've finished making the header for the web app which is a navigation bar. I've made different ones for admins and users. The buttons are not fully functional yet, as we are yet to make the other pages. I've also merged some folders together into the code folder for simplicity. #69

Please review my pull request. Thanks.