Updated nav-bar.html
Added an iframe source code to encapsulate the user-dashboard file into the nav bar. We will be replicating this idea for our other pages as well so that the nav-bar is placed on top of the other pages of our web program.
Team - pls take a look at this change and test user functionality before merging.