Flask Integration and Setup
I have made a .db for our database in order to integrate with flask. I have also made documentation as provided in the pdf file to do this on your own machine.  The first step is to setup a venv and integrate flask, then using sqllite alchemy to complete the 2nd integration. I have also created a template called data_visualizations, for now this is hard coded so you can check functionality.

!! Pls ensure to run this code and the supporting documentation before approving this PR !!