Implemented Test File
In this commit, i implemented test files in the test folder of our repo. The selenium tests are to ensure our web framework is active.

Database Tests: The TestDatabase class checks the existence of specified tables in your SQLite database.

Flask App Tests: The TestFlaskApp class tests the running Flask application. It sends HTTP requests to the base URL and the /data-visualization endpoint.